requirements.txt
Adafruit-Blinka==6.13.0
adafruit-circuitpython-dht==3.5.4
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
backcall==0.2.0
decorator==5.1.0
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
paho-mqtt==1.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pyftdi==0.53.3
Pygments==2.10.0
pyserial==3.5
python-dotenv==0.21.1
pyusb==1.2.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
traitlets==5.1.0
wcwidth==0.2.5
